requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
attrs==25.3.0
ccxt==4.4.82
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
dateparser==1.2.1
frozenlist==1.6.0
idna==3.10
multidict==6.4.3
numpy==2.2.6
pandas==2.2.3
propcache==0.3.1
pycares==4.8.0
pycparser==2.22
pycryptodome==3.23.0
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram==0.19.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
setuptools==80.7.1
six==1.17.0
ta==0.11.0
telegram_text==0.2.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
